LOCS_Energy_Platform_AI/requirement-dev.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mysqlclient==1.4.2
python-dotenv==0.10.1
Werkzeug==0.15.1
